Scoreboard schema, route and sample game data
Added a schema for a scoreboard.  Unfortunately I was getting an error trying to
use the Score type in this schema (something to look into), so I just put in the
score shaped map.

Added an atom called score-board to track changes to the scoreboard.  The atom
contained some sample data for the scoreboard.