Added Player Accounts
Added schemas for:
* PlayerAccount - defines what a player account should look like
* AllPlayerAccounts - defines what collection of all player accounts looks like
* NewPlayer - defines how to request a new player account

Note: Should move the gravatar to the player account as it makes more sense than
attaching it to the score.

Helper functions
Generate new player account and generate a new random UUID for a player-id
value.  Using player-id as the unique value allows players to change their
names.

Routes
* "/new-player" will create a new player account using the helper functions add
  add that player account to the atom.
* "/player-accounts" returns the current player accounts as a vector of maps.