Using Leiningen command to call ring plugin to launch Clojure app
The compojure-api template uses the lein-ring plugin to inject a `-main`
function when run from `lein ring server` command.  So we need to do the same
for Heroku.

If this was a real production system, then we would just add a specific `-main`
class to the namespace where our application should start in, also ensuring that
namespace includes `(:gen-class)` in its `ns` definition.